// src/interface.ts
export enum WsMessageContentTypeEnum {
  txt,
  img,
  video,
}

export enum WsMessageIsShowEnum {
  yes,
  no,
}

export enum WsMessageIsVerifyEnum {
  yes,
  no,
}

export enum WsMessageIsBilibiliEnum {
  yes,
  no,
}

export interface IWsMessage {
  id?: number;
  username?: string;
  origin_username?: string;
  content_type?: WsMessageContentTypeEnum;
  content?: string;
  origin_content?: string;
  redbag_send_id?: number;
  live_room_id?: number;
  user_id?: number;
  ip?: string;
  msg_type?: DanmuMsgTypeEnum;
  user_agent?: string;
  send_msg_time?: number;
  is_show?: WsMessageIsShowEnum;
  is_bilibili?: WsMessageIsBilibiliEnum;
  remark?: string;

  user?: IUser;
  redbag_send?: IRedbagSend;

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}

export interface IRedbagSend {
  id?: number;

  user_id?: number;
  live_room_id?: number;

  total_amount?: string;
  remaining_amount?: string;
  total_nums?: number;
  remaining_nums?: number;
  remark?: string;

  /** 用户信息 */
  user?: IUser;
  /** 直播间信息 */
  live_room?: IGoods;

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export interface IGoods {
  id?: number;
  type?: GoodsTypeEnum;
  name?: string;
  desc?: string;
  short_desc?: string;
  cover?: string;
  price?: number;
  original_price?: number;
  nums?: number;
  pay_nums?: number;
  inventory?: number;
  badge?: string;
  badge_bg?: string;
  remark?: string;

  created_at?: string;
  updated_at?: string;
  deleted_at?: string;
}
// ...
export enum DanmuMsgTypeEnum {
  danmu,
  otherJoin,
  userLeaved,
  system,
  redbag,
  reward,
}
// ...
export interface IWsUserMessage {
  id?: number;
  source_user_id?: number;
  target_user_id?: number;
  ws_user_contact_id?: number;
  content?: string;
  origin_content?: string;
  send_message_time?: number;
  client_env?: string;
  create_at?: string;
  update_at?: string;
  delete_at?: string;
}