// src/types/websocket.ts
export enum WsConnectStatusEnum {
  /** 已连接 */
  connection = 'connection',
  /** 连接中 */
  connecting = 'connecting',
  /** 已连接 */
  connected = 'connected',
  /** 断开连接中 */
  disconnecting = 'disconnecting',
  /** 已断开连接 */
  disconnect = 'disconnect',
  /** 重新连接 */
  reconnect = 'reconnect',
  /** 客户端的已连接 */
  connect = 'connect',
}

/** websocket消息类型 */
export enum WsMsgTypeEnum {
  /** 用户进入聊天 */
  join = 'join',
  /** 用户进入聊天完成 */
  joined = 'joined',
  keepJoined = 'keepJoined',
  /** 用户进入聊天 */
  otherJoin = 'otherJoin',
  /** 用户退出聊天 */
  leave = 'leave',
  /** 用户退出聊天完成 */
  leaved = 'leaved',
  /** 当前所有在线用户 */
  liveUser = 'liveUser',
  /** 用户发送消息 */
  message = 'message',
  /** 房间正在直播 */
  roomLiving = 'roomLiving',
  /** 房间不在直播 */
  roomNoLive = 'roomNoLive',
  /** 获取在线用户 */
  getLiveUser = 'getLiveUser',
  /** 更新加入信息 */
  updateJoinInfo = 'updateJoinInfo',
  /** 心跳 */
  heartbeat = 'heartbeat',
  /** 开始直播 */
  startLive = 'startLive',
  /** 结束直播 */
  endLive = 'endLive',
  /** 直播pk秘钥 */
  livePkKey = 'livePkKey',
  /** 主播禁言用户 */
  disableSpeaking = 'disableSpeaking',
  /** 主播踢掉用户 */
  kick = 'kick',
  keepRtcLiving = 'keepRtcLiving',

  srsOffer = 'srsOffer',
  srsAnswer = 'srsAnswer',
  srsCandidate = 'srsCandidate',

  nativeWebRtcOffer = 'nativeWebRtcOffer',
  nativeWebRtcAnswer = 'nativeWebRtcAnswer',
  nativeWebRtcCandidate = 'nativeWebRtcCandidate',

  msrBlob = 'msrBlob',
  batchSendOffer = 'batchSendOffer',

  changeMaxBitrate = 'changeMaxBitrate',
  changeMaxFramerate = 'changeMaxFramerate',
  changeResolutionRatio = 'changeResolutionRatio',
  changeVideoContentHint = 'changeVideoContentHint',
  changeAudioContentHint = 'changeAudioContentHint',
}

/** 发送消息统一格式 */
export interface IReqWsFormat<T> {
  /** 消息id */
  request_id: string;
  /** 用户socket_id */
  socket_id: string;
  /** 不需要手动传用户代理，从请求头拿 */
  // user_agent: string;
  /** 用户token */
  user_token?: string;
  /** 消息时间戳 */
  time: number;
  data: T;
}

/** 接收消息统一格式 */
export interface IResWsFormat<T> {
  /** 消息id */
  request_id: string;
  /** 消息时间戳 */
  time: number;
  data: T;
}
// ...
export type WsMessageType = IReqWsFormat<IWsMessage>;
export type WsUserMessageType = IReqWsFormat<IWsUserMessage>;
// ...
export type WsOtherJoinType = IResWsFormat<{
  live_room_id: number;
  join_user_info?: IUser;
  join_socket_id: string;
  socket_list: string[];
}>;

/** 开始直播 */
export type WsStartLiveType = IReqWsFormat<{
  type: LiveRoomTypeEnum;
  /** 单位：毫秒 */
  msrDelay: number;
  /** 单位：毫秒 */
  msrMaxDelay: number;
}>;

/** 用户加入直播间 */
export type WsJoinType = IReqWsFormat<{
  live_room_id: number | string;
  cluster_id: string;
  live_room?: ILiveRoom;
  socket_list?: string[];
  duration: number;
  uuid: string;
}>;

/** 用户加入直播间 */
export type WsKeepRtcLivingType = IReqWsFormat<{
  live_room_id: number;
  duration: number;
}>;

/** 用户加入直播间 */
export type WsJoinedType = IResWsFormat<{
  live_room_id?: number;
}>;

/** 用户离开直播间 */
export type WsLeavedType = IResWsFormat<{
  socket_id: string;
  user_info?: IUser;
}>;